use size_t instead of int. -Wsign-compare

+++ inst/include/tools/DataDots.h
@@ -48,7 +48,7 @@ namespace Rcpp {
         inline bool single_env() const{
             if( environments.size() < 2 ) return true ;
             SEXP first = environments[0] ;
-            for( int i=1; i<environments.size(); i++){
+            for( size_t i=1; i<environments.size(); i++){
                 if( first != environments[i] ) return false ;    
             }
             return true ;
